+++ R/bracl.R
@@ -315,6 +315,8 @@ print.summary.bracl <- function(x, digits = x$digits, ...) {
     cat("\nResidual Deviance:", format(x$deviance), "\n")
     cat("Log-likelihood:", format(x$logLik), "\n")
     cat("AIC:", format(x$AIC), "\n")
+    cat("\n\nType of estimator:", x$type, get_type_description(x$type))
+    cat("\n", "Number of Fisher Scoring iterations: ", x$iter, "\n", sep = "")
     if (!is.null(correl <- x$correlation)) {
         p <- dim(correl)[2L]
         if (p > 1) {