+++ R/brmultinom.R
@@ -393,7 +393,9 @@ print.summary.brmultinom <- function(x, digits = x$digits, ...)
     }
     cat("\nResidual Deviance:", format(x$deviance), "\n")
     cat("Log-likelihood:", format(x$logLik), "\n")
-    cat("AIC:", format(x$AIC), "\n")
+    cat("AIC:", format(x$AIC))
+    cat("\n\nType of estimator:", x$type, get_type_description(x$type))
+    cat("\n", "Number of Fisher Scoring iterations: ", x$iter, "\n", sep = "")
     if (!is.null(correl <- x$correlation)) {
         p <- dim(correl)[2L]
         if (p > 1) {